requirements.txt
flask==3.0.0
pandas==2.1.4
openai==1.6.1
pyyaml==6.0.1
tenacity==8.2.3